continuing cleanup
